fixed bug in pseudorandom number implict type casting

+++ psrmodels/time_dependent/ConvGenDistribution.py
@@ -172,7 +172,7 @@ class ConvGenDistribution(object):
             raise Exception("Some state sets do not match the shape of corresponding transition matrix")
 
       # set output array
-      output = np.ascontiguousarray(np.empty((n_sim,timesteps_in_season)),dtype=np.float32)
+      output = np.ascontiguousarray(np.empty((n_sim,timesteps_in_season),dtype=np.float32),dtype=np.float32)
 
       #print("output shape: {s}".format(s=output.shape))
       #print("output before: {o}".format(o=output))
